% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copomscenarios-class.R
\docType{class}
\name{COPOMScenarios-class}
\alias{COPOMScenarios-class}
\title{COPOMScenarios class}
\description{
COPOMScenarios class is an interpolation that uses scenarios for spot rates
after COPOM meetings together with flat forward COPOM interpolation.
}
\details{
The scenarios can be defined in two ways:

- future values for SELIC rates
- future values for changes in SELIC rates

both associated to each meeting.
}
\seealso{
FlatForwardCOPOM

interp_flatforwardcopom
}
