% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatforwardcopom-class.R
\docType{class}
\name{FlatForwardCOPOM-class}
\alias{FlatForwardCOPOM-class}
\title{FlatForwardCOPOM interpolation}
\description{
FlatForwardCOPOM interpolation extends Interpolation class from `fixedincome`
package.
}
\details{
This interpolation considers that the interest rate between COPOM meetings
in Brazil are flat, instead of being flat betweet bonds maturities.
This is mainly relevant for the short term of the term structure where
the interpolation is used to price private bonds and interest rate
derivatives.
This is discussed in the book Brazilian Derivatives and Securities.
}
\references{
Marcos C. S. Carreira and Richard J. Brostowicz.
            Brazilian Derivatives and Securities, Palgrave Macmillan, 2016
}
