% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation-copom-class.R
\name{get_copom_dates}
\alias{get_copom_dates}
\title{Historical COPOM dates}
\usage{
get_copom_dates()
}
\value{
A vector with `Date` objects.
}
\description{
Historical COPOM dates since 2015.
}
